User Story: As a user, I should be able to updated and login files renamed
